Refine readability
