+++ crates/slipstream-client/src/error.rs
@@ -1,0 +1,22 @@
+use std::fmt;
+
+#[derive(Debug)]
+pub struct ClientError {
+    message: String,
+}
+
+impl ClientError {
+    pub(crate) fn new(message: impl Into<String>) -> Self {
+        Self {
+            message: message.into(),
+        }
+    }
+}
+
+impl fmt::Display for ClientError {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(f, "{}", self.message)
+    }
+}
+
+impl std::error::Error for ClientError {}
