+++ crates/slipstream-client/src/runtime/setup.rs
@@ -1,0 +1,27 @@
+use crate::error::ClientError;
+use std::net::{Ipv6Addr, SocketAddr, SocketAddrV6};
+use tokio::net::UdpSocket as TokioUdpSocket;
+
+pub(crate) fn compute_mtu(domain_len: usize) -> Result<u32, ClientError> {
+    if domain_len >= 240 {
+        return Err(ClientError::new(
+            "Domain name is too long for DNS transport",
+        ));
+    }
+    let mtu = ((240.0 - domain_len as f64) / 1.6) as u32;
+    if mtu == 0 {
+        return Err(ClientError::new(
+            "MTU computed to zero; check domain length",
+        ));
+    }
+    Ok(mtu)
+}
+
+pub(crate) async fn bind_udp_socket() -> Result<TokioUdpSocket, ClientError> {
+    let bind_addr = SocketAddr::V6(SocketAddrV6::new(Ipv6Addr::UNSPECIFIED, 0, 0, 0));
+    TokioUdpSocket::bind(bind_addr).await.map_err(map_io)
+}
+
+pub(crate) fn map_io(err: std::io::Error) -> ClientError {
+    ClientError::new(err.to_string())
+}
